libdwarf: Fix a potential memory leak.
'dwarf_get_abbrev()' can fail whilst in the middle of constructing
a 'Dwarf_Abbrev' instance, leaking the memory for the partially
constructed instance.

This change refactors the allocation and release code for 'struct
_Dwarf_Abbrev' structures into independently callable helper
functions, and then uses these to ensure that any partially
constructed structures are destroyed in the event of an error.

Reported by: Coverity Scan
Coverity CID: 352394


git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@4046 95820547-d848-0410-985e-9ae8fe0fa350