elfconstants: remove DT_MAXPOSTAGS.
This symbol holds the number of small-valued DT_* tags known
at compile time.

However, ELF files in the wild could contain DT_* tags that
are not known at compile time and/or could contain known
but non-small-valued tags in the range DT_LOOS (0x6000000D)
to DT_HIPROC (0x7FFFFFFF).  ELF-handling tools are expected
to handle such tags gracefully, which means that the value of
DT_MAXPOSTAGS is not useful in practice.

Remove the symbol to reduce maintenance overhead.


git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@4259 95820547-d848-0410-985e-9ae8fe0fa350