tools/netbsd-base-system-import.sh: Suppress unwanted expansion of $NetBSD$.
Before this change, instances of the literal string '$NetBSD$' within
this script were expanded by NetBSD's CVS if this script was imported
into NetBSD's sources and then checked out.  When subsequently used,
the script would then copy the expanded string into the locations
in the files being imported where the literal string '$NetBSD$'
was desired.

Although harmless (as CVS would eventually use the correct expansion
when the imported files were checked out later), the use of the
expanded value for '$NetBSD$' caused confusing diffs prior to check-in.

With this change, the literal string '$NetBSD$' should be placed in
the desired locations as-is.


git-svn-id: svn://svn.code.sf.net/p/elftoolchain/code/trunk@4326 95820547-d848-0410-985e-9ae8fe0fa350